Refactored gens module
